+++ foxes/output/__init__.py
@@ -5,9 +5,6 @@ Output tools and functions.
 from .output import Output as Output
 from .farm_layout import FarmLayoutOutput as FarmLayoutOutput
 from .farm_results_eval import FarmResultsEval as FarmResultsEval
-from .rose_plot import RosePlotOutput as RosePlotOutput
-from .rose_plot import StatesRosePlotOutput as StatesRosePlotOutput
-from .rose_plot import WindRoseBinPlot as WindRoseBinPlot
 from .results_writer import ResultsWriter as ResultsWriter
 from .state_turbine_map import StateTurbineMap as StateTurbineMap
 from .turbine_type_curves import TurbineTypeCurves as TurbineTypeCurves
@@ -18,6 +15,11 @@ from .slices_data import SlicesData as SlicesData
 from .rotor_point_plots import RotorPointPlot as RotorPointPlot
 from .state_turbine_table import StateTurbineTable as StateTurbineTable
 from .plt import plt as plt
+from .multiple_farms import MultipleFarmsOutput as MultipleFarmsOutput
+
+from .rose_plot import RosePlotOutput as RosePlotOutput
+from .rose_plot import StatesRosePlotOutput as StatesRosePlotOutput
+from .rose_plot import WindRoseBinPlot as WindRoseBinPlot
 
 from .flow_plots_2d import FlowPlots2D as FlowPlots2D
 from .flow_plots_2d import write_chunk_ani_xy as write_chunk_ani_xy